此次提交没有实现更多的功能点，只是针对前后端连携的问题做了一些修改工作，此次提交的主要目的是和另一位后端做合并分支操作

+++ src/main/java/com/secondhandbookstore/pojo/User.java
@@ -14,7 +14,7 @@ public class User {
     private Boolean manage;
     private String img;
     Integer gender;
-    private Float blance;
+    private Double blance;
     private Integer area;
     private String address;
     private String phone;
